dev-ruby/thor: cleanup ruby27-only USE_RUBY for tests
We have a hack in dev-ruby/thor to avoid too many test dependencies
when porting to a newer Ruby, which is fine (we do it for Python
all the time too), but in this case, it only contains now-removed
(or about to be removed) Rubies.

After 92902571781915a800816c1a89e68c3d1b9cad94, this is no longer
ignored when we try to cleanup/mask/disable old Rubies.

Upon reflection, I think this is the right behaviour:
1. If real runtime dependencies are only needed for older Rubies
then we want to clean up the condition once it's useless, not ignore it;

2. If test conditions are only used for old rubies, we either want
to clean it up (because it's obsolete) or fix it (if we left a trick
behind when bootstrapping a new Ruby, as it means we're not testing it).

While at it, add a comment explaining what we're doing, as it's a
useful thing to know about.

Bonus: in this case, it was obscuring a real issue - we weren't
running tests for >=ruby30 long after ruby30 got added and we missed
that the hash tests failed.