sys-devel/binutils: add last-minute sanity check (WIP)
Takes inspiration from sys-libs/glibc's sanity check as well as the one
I recently added for sys-libs/musl (c1078bc34bba21a2ff31289744e17f5addc07c89).

The idea is to just check that the new binutils (gas, ld) produce working
binaries for a trivial example. PR32991 is a nice example of where this
could help find a problem.

This doesn't help with all cases, but it is somewhat related to the linked
Gentoo bugs.

It's more important for the case where the user is on an odd target or has
specified some unusual flags and binutils is completely broken to avoid
merging and bricking the machine (as a fixed binutils won't be buildable).

Better safe than sorry.

Future work:
* Check all ABIs (like -m32)
* Do the same for gcc?

Bug: https://bugs.gentoo.org/622690
Bug: https://bugs.gentoo.org/802036
Bug: https://sourceware.org/PR32991
Signed-off-by: Sam James <sam@gentoo.org>